Move config related files to it's own folder
