+++ atlas/core/config/ConfigNotification.hpp
