+++ atlas/core/import/GameImportData.hpp
@@ -9,7 +9,7 @@
 
 #include <filesystem>
 
-#include "core/config.hpp"
+#include "core/config/config.hpp"
 #include "core/database/record/GameData.hpp"
 #include "core/database/record/game/Game.hpp"
 #include "core/gamelist/utils.hpp"
