+++ atlas/ui/delegates/RecordBannerDelegate.hpp
@@ -8,7 +8,7 @@
 #include <QAbstractItemDelegate>
 #include <QtCore/QModelIndex>
 
-#include "core/config.hpp"
+#include "core/config/config.hpp"
 #include "core/database/record/game/Game.hpp"
 #include "core/utils/QImageBlur.hpp"
 class RecordListModel;
