+++ atlas/ui/importer/simpleImporter/SIModel.hpp
@@ -8,7 +8,7 @@
 #include <QAbstractItemModel>
 #include <QDir>
 
-#include "core/config.hpp"
+#include "core/config/config.hpp"
 #include "core/logging/logging.hpp"
 
 enum class SupportingType
